Refactor calculating view numbers
Move related view config numbers to ViewConfig, and refactor usage so
key numbers are in the `ViewConfig` type while others are derived from
this.

The size of the SVG element is now the field others are derived from
rather than the other way round, meaning the single `scale` value can
now be changed to change the number of cells drawn in the same-sized SVG
element (required to implement zooming in/out).