Setup compiling elm-css with Webpack loader
Due to running into limitations with inline styles (lack of being able
to use selectors like `:hover` or `.something + .something`) I want to
migrate to generating a stylesheet with `elm-css` rather than using
inline styles. This is the first stage of this, using
`elm-css-webpack-loader` to load some initial styles; this means the
existing CSS file is now entirely unneeded as well.